Renaming checkbox mixin
