+++ addon/mixins/checkbox.js
